\name{cmap}
\alias{cmap}
\docType{data}
\title{
Data from Connectivity map build 01
}
\description{
It is the sample data used for testing the function RGSEAsd.
}
\usage{data(cmap)}
\format{
  The format is:
 num [1:22268, 1:6] 0.4892 -0.6137 3.5242 -0.0139 -2.0255 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:22268] "1007_s_at" "1053_at" "117_at" "121_at" ...
  ..$ : chr [1:6] "thioridazine" "tretinoin" "prochlorperazine" "chlorpromazine" ...
}
\details{
The query data is the instance 5202764005791175120104.C08, treated with thioridazine.

The reference data is 5202764005789148112904.G05,	X5202764005789148112904.F03,	X5202764005789148112904.F05,	X5202764005789148112904.E02,	X5202764005789148112904.E04.
They were treated with tretinoin,	prochlorperazine,	chlorpromazine,	vorinostat,	sirolimus respectively.
All the data were generated by MCF7 cell line.
}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5258
}
\references{
Lamb J, Crawford ED, Peck D, Modell JW et al. The Connectivity Map: using gene-expression signatures to connect small molecules, genes, and disease. Science 2006 Sep 29;313(5795):1929-35. PMID: 17008526
}
\examples{
data(cmap)
}
\keyword{datasets}
