\name{e1}
\alias{e1}
\docType{data}
\title{
Data from GDS4102
}
\description{
This is the query data for testing the function RGSEAfix.
}
\usage{data(e1)}
\format{
  The format is:
 num [1:54675, 1:2] 1012 44.32 43.03 36.65 4.92 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:54675] "1007_s_at" "1053_at" "117_at" "121_at" ...
  ..$ : chr [1:2] "tumor" "normal"
}
\details{
The two data are GSM414924 GSM414975 repectively.
}
\source{
http://www.ncbi.nlm.nih.gov/sites/GDSbrowser?acc=GDS4102
}
\references{
Pei H, Li L, Fridley BL, Jenkins GD et al. FKBP51 affects cancer cell response to chemotherapy by negatively regulating Akt. Cancer Cell 2009 Sep 8;16(3):259-66. PMID: 19732725
}
\examples{
data(e1)
}
\keyword{datasets}
