\name{e2}
\alias{e2}
\docType{data}
\title{
Data from GDS4100
}
\description{
This is the reference data for testing the function RGSEAfix.
}
\usage{data(e2)}
\format{
  The format is:
 num [1:54675, 1:24] 1.48 1.19 0.67 2.75 NA 0.51 1.68 NA NA 1.99 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:54675] "1007_s_at" "1053_at" "117_at" "121_at" ...
  ..$ : chr [1:24] "tumor" "tumor" "tumor" "tumor" ...
}
\details{
This dataset contains all the 4 data of GDS4100.
}
\source{
http://www.ncbi.nlm.nih.gov/sites/GDSbrowser?acc=GDS4100
}
\references{
Zhang L, Farrell JJ, Zhou H, Elashoff D et al. Salivary transcriptomic biomarkers for detection of resectable pancreatic cancer. Gastroenterology 2010 Mar;138(3):949-57.e1-7. PMID: 19931263
}
\examples{
data(e2)
}
\keyword{datasets}
