\name{RGSEApredict}
\alias{RGSEApredict}
\title{
Predict the class of the query data with the result of RGSEA functions
}
\description{
Predict the class of the query data with the result of RGSEA functions--RGSEAfix or RGSEAsd
}
\usage{
RGSEApredict(RGSEAresult, refclasses)
}
\arguments{
  \item{RGSEAresult}{
The first item of the results generated by RGSEA functions. 
}
  \item{refclasses}{
A character vector. The classes of the reference data.
}
}
\author{
Chengcheng Ma
}
\examples{
if(interactive()) {
    data(e1)
    data(e2)
    RGSEAfix(e1,e2, queryclasses=colnames(e1), refclasses=colnames(e2),   
random=20000, featurenum=1000, iteration=100)->test
    RGSEApredict(test[[1]], colnames(e2))
}
}
